[Binding][iCloud] Type of setPrefsUpdateTime in ServerContext changed from int to long to solve deserialization problems where the number exceeds int.

+++ addons/binding/org.openhab.binding.icloud/src/main/java/org/openhab/binding/icloud/internal/json/iCloud/ServerContext.java
@@ -189,13 +189,13 @@ public class ServerContext {
         this.info = info;
     }
 
-    private int prefsUpdateTime;
+    private long prefsUpdateTime;
 
-    public int getPrefsUpdateTime() {
+    public long getPrefsUpdateTime() {
         return this.prefsUpdateTime;
     }
 
-    public void setPrefsUpdateTime(int prefsUpdateTime) {
+    public void setPrefsUpdateTime(long prefsUpdateTime) {
         this.prefsUpdateTime = prefsUpdateTime;
     }
